refactor(tocco-ui): replace loading component
- remove ToccoIcon and use classic spinner instead.
