--- packages/tocco-ui/src/IconTocco/index.js
@@ -1,8 +1,0 @@
-import IconTocco from './IconTocco'
-import {StyledIconToccoSvg, StyledIconToccoWrapper} from './StyledIconTocco'
-
-export {
-  IconTocco as default,
-  StyledIconToccoSvg,
-  StyledIconToccoWrapper
-}
