+++ packages/tocco-ui/src/LoadingSpinner/index.js
@@ -1,0 +1,5 @@
+import LazyLoadingSpinner from './LazyLoadingSpinner'
+
+export {
+  LazyLoadingSpinner as default
+}
