+++ packages/tocco-ui/src/Select/LoadingIndicator.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import IconTocco from '../IconTocco'
+import IconTocco from '../LoadingSpinner'
 
 const LoadingIndicator = () => <span><IconTocco size="1.8rem"/></span>
